pyproject.toml
[project]
name = "react-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.6",
    "python-dotenv>=1.2.2",
]
